+++ Backend/Sprint_Manager/src/main/java/com/app/dao/Employee_dao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.app.entities.Employee;
 import com.app.entities.Sprint;
 
-public interface Employee_dao extends JpaRepository<Employee,Long> {
+public interface Employee_dao extends JpaRepository<Employee,Integer> {
 
 }
