+++ Backend/Sprint_Manager/src/main/java/com/app/dao/ManagerDao.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 
-public interface ManagerDao extends JpaRepository<Manager,Long> {
+public interface ManagerDao extends JpaRepository<Manager,Integer> {
 	
 	Optional<Manager> findByEmail(String email);
