+++ Backend/Sprint_Manager/src/main/java/com/app/dao/Sprint_entries_dao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.app.entities.Sprint;
 import com.app.entities.Sprint_entry;
 
-public interface Sprint_entries_dao extends JpaRepository<Sprint_entry,Long> {
+public interface Sprint_entries_dao extends JpaRepository<Sprint_entry,Integer> {
 
 }
