+++ Backend/Sprint_Manager/src/main/java/com/app/service/Sprint_service.java
@@ -6,9 +6,9 @@ import com.app.entities.Sprint;
 
 public interface Sprint_service {
 
-	public List<Sprint> getAllSprints(int id);
+	public List<Sprint> getAllSprints(Integer id);
 	public Sprint addNewEmp(Sprint sp);
 	public Sprint updateEmpDetails(Sprint sp);
-	public String deleteEmpDetails(Long sprintId);
+	public String deleteEmpDetails(Integer sprintId);
 	
 }
